requirements.txt
streamlit
tensorflow-cpu
numpy
opencv-python-headless
matplotlib